+++ tests/controllers/Base_Controller_Test.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace WP_Swapper\Tests;
+namespace WP_Swapper\Tests\Controllers;
+
+use WP_Swapper\Tests\TestCase;
 
 /**
 * abstract base class for teseting controllers
