+++ tests/controllers/Header_Controller_Test.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace WP_Swapper\Tests;
+namespace WP_Swapper\Tests\Controllers;
 
-use WP_Swapper\Controllers\Footer_Controller;
+use WP_Swapper\Controllers\Header_Controller;
 
 /**
-* Class for testing the Footer Controller
+* Class for testing the Header Controller
 *
 * @since 0.0.1
 */
-class Footer_Controller_Test extends Base_Controller_Test {
+class Header_Controller_Test extends Base_Controller_Test {
     /**
     * The Initial content
     *
@@ -17,7 +17,7 @@ class Footer_Controller_Test extends Base_Controller_Test {
     *
     * @var string
     */
-    protected $initial_content = '<footer>Initial Content</footer>';
+    protected $initial_content = '<header>Initial Content</header>';
 
     /**
     * The new content
@@ -26,7 +26,7 @@ class Footer_Controller_Test extends Base_Controller_Test {
     *
     * @var string
     */
-    protected $new_content = '<footer>Updated Content</footer>';
+    protected $new_content = '<header>Updated Content</header>';
 
     /**
     * The component name
@@ -35,32 +35,32 @@ class Footer_Controller_Test extends Base_Controller_Test {
     *
     * @var string used in the header
     */
-    protected $component_name = 'Footer';
+    protected $component_name = 'Header';
 
     /**
-    * Provide the Footer_Controller class for testing
+    * Provide the Header_Controller class for testing
     *
     * @since 0.0.1
     *
     * @param string $content
     *
-    * @return Footer_Controller
+    * @return Header_Controller
     */
     protected function get_controller($content) {
-        return new Footer_Controller($content);
+        return new Header_Controller($content);
     }
     /**
-    * Test the returned view of footer component
+    * Test the returned view of header component
     *
     * @since 0.0.1
     *
     * @runInSeparateProcess
     */
     public function testView() {
-        $controller = new Footer_Controller($this->initial_content);
+        $controller = new Header_Controller($this->initial_content);
 
         if ($controller->render() !== '') {
-            $expectedId = 'id="changed-footer"';
+            $expectedId = 'id="changed-header"';
             $this->assertStringContainsString($expectedId, $controller->render(), 'View should contain the correct Id');
         }
     }
